Convert the summary and top_domains tests into unit tests
